extract cache to file
